+++ signify/authenticode/signed_file/__init__.py
@@ -1,5 +1,6 @@
 from .base import AuthenticodeFile
 from .ctl import CtlFile
+from .flat import FlatFile
 from .pe import SignedPEFile, SignedPEFingerprinter
 from .signed_data import AuthenticodeSignedDataFile
 
@@ -7,6 +8,7 @@ __all__ = [
     "AuthenticodeFile",
     "AuthenticodeSignedDataFile",
     "CtlFile",
+    "FlatFile",
     "SignedPEFile",
     "SignedPEFingerprinter",
 ]